Adds The Great Eye of Sauron
I'm *still* not entirely happy with this. It tells me a lot about the
things that I'm not that interested in (tweets that aren't replies or
DMs) and not enough of people in slack channels that I'm interested in
or other things. I really don't know which way to go with this or what
might be a better alternative.