Improves agenda view. Needs improvement.
This *almost* gives me the focused agenda view that I want with tasks, a
quick view over what I've done. What I have to do today and and if there
are any big things I need to prepare for tomorrow. The issue is that
yesterday and tomorrow and *really* the previous and next working days
and I should improve the code to handle things like bank holidays.